\encoding{UTF-8}
\name{adephylo-package}
\alias{adephylo-package}
\alias{adephylo}
\docType{package}
\title{The adephylo package}

\description{This package is devoted to exploratory analysis of
  phylogenetic comparative data.
 
  
 

   
   === MANIPULATING DATA ===\cr
   Several functions allow one to manipulate \linkS4class{genind} or
   \linkS4class{genpop} objects\cr
   - \code{\link{}}: convert a \linkS4class{genind} object
   to a  \linkS4class{genpop} \cr
  

   === ANALYZING DATA ===\cr
   :\cr
   - \code{\link{}}: performs HWE test for all
 
   
   To cite adephylo, please use the reference given by
   \code{citation("adephylo")}.
 }
\details{
  \tabular{ll}{
    Package: \tab adephylo\cr
    Type: \tab Package\cr
    Version: \tab 1.0-0\cr
    Date: \tab 2008-12-30 \cr
    License: \tab GPL (>=2)
  }  
}
\author{Thibaut Jombart <jombart@biomserv.univ-lyon1.fr>\cr
with contributions Stéphane Dray}
\keyword{manip}
\keyword{multivariate}
\seealso{The \code{ade4} package for multivariate analysis}
